// src/enums/profitability.enums.ts
export enum ProfitabilityTableTypeEnum {
  PRODUCTS = 'Products_Table',
  ORDERS = 'Orders_Table',
  AMAZON_ORDERS = 'Amazon_Orders_Table',
  AMAZON_PRODUCTS = 'Amazon_Products_Table',
}

export enum ProfitabilityTableTitlesEnum {
  PROFITABILITY_HOME = 'Profitability_Home',
  PROFITABILITY_TRENDS = 'Profitability_Trends',
  AMZ_PROFITABILITY_TRENDS = 'Amazon_Profitability_Trends',
  PROFITABILITY_PNL = 'Profitability_PnL',
  PROFITABILITY_PNL_ORDERS = 'Profitability_PnL_Orders',
  PROFITABILITY_PNL_PRODUCTS = 'Profitability_PnL_Products',
  AMZ_PROFITABILITY_PNL_ORDERS = 'Amazon_Profitability_PnL_Orders',
  AMZ_PROFITABILITY_PNL_PRODUCTS = 'Amazon_Profitability_PnL_Products',
}